add external url and image to card
Since this is an external image we should use <a> and not <Link>
But for whatever reason I kept getting a broken url when passing it
as a prop to the <a> and <source> tags, using .join('') fixed it.
But I have no idea why it came all broke apart.

We should dive deeper into this, but we can look at it after the
presentation this sunday.