Feat/24 Create results static page with one result (#27)
* add rough content to the page

* create card component

I created a components folder, created a card component and moved
all the card code there. At the moment it's getting it's styles from results,
we should probably a separate card.css and move those styles there.

* make the card content dynamic, work on the mapping logic

Now we are mapping through the results, BUT I need to fix
the urls, and, ideally, keep the object only in the separate file
and import it

* add external url and image to card

Since this is an external image we should use <a> and not <Link>
But for whatever reason I kept getting a broken url when passing it
as a prop to the <a> and <source> tags, using .join('') fixed it.
But I have no idea why it came all broke apart.

We should dive deeper into this, but we can look at it after the
presentation this sunday.

* import results data

Instead of keeping the data inside the results page, we now are
importing it from a separate file. For some reason it hadn't worked
before, my guess, I wrote something wrong because I was sleepy.

* add different images

I got a few temporary different images just so we could have
a better feel for what the results page looks like. The alt text
is still missing, but I don't think we should worry about that now.